# Optimize `HashCode.AddBytes` for inputs larger than 16 bytes.